API Reference

Models

AeroPanels.AeroModel2D — Type

struct AeroModel2D{T<:Real} <: AeroModel

A 2D aerodynamic model representing a collection of lifting surfaces and their steady wake.

mesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T<:Real, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{Point3{T}, 1})}

ringMesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T<:Real, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{Point3{T}, 1})}

wakeMesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T<:Real, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{Point3{T}, 1})}

sizes::Sizes

wakeSizes::Sizes

panelProperties::PanelProperties{T} where T<:Real

modelProperties::AeroModelProperties{T} where T<:Real

AIC::LinearAlgebra.LU{T, Matrix{T}, Vector{Int64}} where T<:Real

segmentProps::SegmentProperties{T} where T<:Real

controlSurfaces::Array{AeroPanels.ControlSurface{T}, 1} where T<:Real

monitorPoints::Array{AeroPanels.MonitorPoint{T}, 1} where T<:Real

AeroPanels.UnsteadyAeroModel2D — Type

struct UnsteadyAeroModel2D{T} <: AeroModel

An unsteady 2D aerodynamic model based on the Continuous-Time Unsteady Vortex Lattice Method.

mesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{Point3{T}, 1})}

ringMesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{Point3{T}, 1})}

wakeMesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{Point3{T}, 1})}

sizes::Sizes

wakeSizes::Sizes

panelProperties::PanelProperties

modelProperties::AeroModelProperties

segmentProps::SegmentProperties

K8::SparseArrays.SparseMatrixCSC{T, Int64} where T

K9::SparseArrays.SparseMatrixCSC{T, Int64} where T

L3::Matrix

L4::Matrix

L5::Matrix

L6::Matrix

Γw0Indices::Vector{Int64}

Γw1Indices::Vector{Int64}

ΓbTEIndices::Vector{Int64}

controlSurfaces::Array{AeroPanels.ControlSurface{T}, 1} where T

monitorPoints::Array{AeroPanels.MonitorPoint{T}, 1} where T

AeroPanels.AeroModelProperties — Type

struct AeroModelProperties{T}

Aerodynamic model properties.

c::Any

b::Any

S::Any

symmXZ::Bool

symmXY::Bool

bodyFixedCS::SMatrix{3, 3, T, 9} where T

CG::SVector{3}

AeroPanels.PanelProperties — Type

struct PanelProperties{T}

Properties of the panels in the aerodynamic mesh.

normal::Array{Point3{T}, 1} where T

rMid::Array{Point3{T}, 1} where T

rCollocation::Array{Point3{T}, 1} where T

rControl::Array{Point3{T}, 1} where T

area::Vector

span::Vector

Solvers

AeroPanels.AeroSolve — Function

AeroSolve(vb, model; kwargs...)

Solve the steady aerodynamic problem for a given body velocity vb. Returns a cache containing all circulations and forces.

AeroSolve(model::UnsteadyAeroModel2D, Γw1_0, tspan, input_func!; solver)

Solve an unsteady time-domain simulation and auto-compute forces/coefficients.

AeroPanels.AeroSolve! — Function

AeroSolve!(cache, vb, ωb, δc, dδc, rs, vs, model)
AeroSolve!(cache, vb, ωb, δc, dδc, rs, vs, model, ρ)

In-place version of AeroSolve.

AeroSolve!(
    cache,
    vb,
    ab,
    ωb,
    dωb,
    δc,
    dδc,
    ddδc,
    rs,
    vs,
    as,
    Γw1,
    model
)
AeroSolve!(
    cache,
    vb,
    ab,
    ωb,
    dωb,
    δc,
    dδc,
    ddδc,
    rs,
    vs,
    as,
    Γw1,
    model,
    ρ
)

In-place version of the unsteady AeroSolve.

AeroPanels.CreateCacheArrays — Function

CreateCacheArrays(model, T)

Create and return a tuple of cache arrays for the steady solver.

CreateCacheArrays(model, T)

Create and return a tuple of cache arrays for the unsteady solver.

AeroPanels.GetStabilityCoefficients — Function

GetStabilityCoefficients(cache, vb, ρ, model)

Return the stability axis force and moment coefficients.

AeroPanels.GetTotalForces — Function

GetTotalForces(cache, model)

Return the total force and moment at the reference point (CG).

AeroPanels.AddSteadyKinematics! — Function

AddSteadyKinematics!(
    rVertex,
    vVertex,
    vb,
    ωb,
    δc,
    dδc,
    rs,
    vs,
    model
)

Calculate the deformed mesh coordinates and the kinematic velocity at the mesh vertices.

AeroPanels.AddUnsteadyKinematics! — Function

AddUnsteadyKinematics!(
    aVertex,
    rVertex,
    ab,
    dωb,
    ddδc,
    as,
    model
)

Calculate the unsteady kinematics (vertex accelerations).

AeroPanels.CalculateNormalwash! — Function

CalculateNormalwash!(b, rVertex, vVertex, model)

Calculate the normal wash on each panel.

AeroPanels.CalculateDNormalwash! — Function

CalculateDNormalwash!(db, rVertex, vVertex, aVertex, model)

Calculate the time derivative of the normal wash vector db.

AeroPanels.SolveCirculation! — Function

SolveCirculation!(Γp, Γw, Γs, b, model)

Solve for panel, wake, and segment circulations in-place.

AeroPanels.SolveUnsteadyCirculation! — Function

SolveUnsteadyCirculation!(
    dΓw1,
    Γw1,
    dΓb,
    Γb,
    Γw,
    Γs,
    b,
    db,
    V,
    model
)

Solve for the time derivatives and intermediate circulations of the unsteady system.

AeroPanels.CalculateAerodynamicForce! — Function

CalculateAerodynamicForce!(
    Fa,
    Γp,
    Γw,
    Γs,
    vVertex,
    model,
    ρ
)

Calculate aerodynamic forces on segments in-place into vector Fa.

AeroPanels.CalculateUnsteadyAerodynamicForce! — Function

CalculateUnsteadyAerodynamicForce!(
    Fa,
    dΓb,
    Γb,
    Γw,
    Γs,
    vVertex,
    model,
    ρ
)

Calculate aerodynamic forces for the unsteady model.

AeroPanels.MonitorPointLoads! — Function

MonitorPointLoads!(Fmp, cache, model)

Calculate aerodynamic loads at monitor points.

Secondary Functions

AeroPanels.AeroInputs — Type

mutable struct AeroInputs{T}

A struct holding the time-dependent inputs for unsteady aerodynamic simulations. These inputs are meant to be updated in-place during time integration.

AeroPanels.AeroSurface2D — Type

struct AeroSurface2D{T<:Real} <: AeroSurface{T<:Real}

A 2D aerodynamic surface defined by a mesh of points in 3D space.

X::Matrix{T} where T<:Real

Y::Matrix{T} where T<:Real

Z::Matrix{T} where T<:Real

nc::Int64

ns::Int64

AeroPanels.Mirror — Function

Mirror(surface, axis)

Mirror an AeroSurface2D along a given axis (1=X, 2=Y, 3=Z).

AeroPanels.SegmentProperties — Type

struct SegmentProperties{T}

Properties and mapping indices for panel segments.

mid::Array{Point3{T}, 1} where T

r::Array{Point3{T}, 1} where T

aic3Ring::Array{Point3{T}, 2} where T

aic3Wake::Array{Point3{T}, 2} where T

indPosSeg::Vector{Int64}

indPosPanel::Vector{Int64}

indNegSeg::Vector{Int64}

indNegPanel::Vector{Int64}

nSpanSegments::Int64

nChordSegments::Int64

nTotalSegments::Int64

AeroPanels.SegmentCirculation — Function

SegmentCirculation(Γp, sp)

Return the segment circulation vector.

AeroPanels.SegmentCirculation! — Function

SegmentCirculation!(Γseg, Γp, sp)

Calculate segment circulations from panel circulations in-place.

AeroPanels.ProcessSegments — Function

ProcessSegments(meshRing, sizesRing, meshWake, sizesWake, symmXZ)

Initialize and calculate all segment properties.

AeroPanels.NumberOfStates — Function

NumberOfStates(model)

Return the number of states in the unsteady state-space model.

AeroPanels.UnsteadyWakeInfluence — Function

UnsteadyWakeInfluence(
    rCollocation,
    normal,
    ringMesh,
    wakeMesh,
    bodySizes,
    wakeSizes,
    symmXZ,
    Δxw
)

Calculate the unsteady aerodynamic influence coefficients and state-space matrices. Based on Binder (2017) and AIAA (2018) publications.

Indexing and Utility

AeroPanels.Sizes — Type

struct Sizes

Struct to store the sizes and pre-calculated indices of the mesh elements.

sizes::Vector{Tuple{Int64, Int64}}

nSurfaces::Int64

totalPanels::Int64

totalVertices::Int64

totalSpanSegments::Int64

totalChordSegments::Int64

panelIndices::Vector{Matrix{Int64}}

vertexIndices::Vector{Matrix{Int64}}

spanSegmentIndices::Vector{Matrix{Int64}}

chordSegmentIndices::Vector{Matrix{Int64}}

AeroPanels.TEPanelIndex — Function

TEPanelIndex(sizes::Sizes)

Return a vector of global indices for all panels along the trailing edges.

AeroPanels.LEPanelIndex — Function

LEPanelIndex(sizes::Sizes)

Return a vector of global indices for all panels along the leading edges.

AeroPanels.PanelIndex — Function

PanelIndex(s::Int, i::Int, j::Int, sizes::Sizes)

Return the global index of the panel (i, j) on surface s.