API Reference

Models

AeroPanels.AeroModel2D — Type

struct AeroModel2D{T<:Real} <: AeroModel

A 2D aerodynamic model representing a collection of lifting surfaces and their steady wake.

mesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T<:Real, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{GeometryBasics.Point3{T}, 1})}

ringMesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T<:Real, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{GeometryBasics.Point3{T}, 1})}

wakeMesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T<:Real, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{GeometryBasics.Point3{T}, 1})}

sizes::Sizes

wakeSizes::Sizes

panelProperties::PanelProperties{T} where T<:Real

modelProperties::AeroModelProperties{T} where T<:Real

AIC::LinearAlgebra.LU{T, Matrix{T}, Vector{Int64}} where T<:Real

segmentProps::SegmentProperties{T} where T<:Real

AeroPanels.UnsteadyAeroModel2D — Type

struct UnsteadyAeroModel2D{T} <: AeroModel

An unsteady 2D aerodynamic model based on the Continuous-Time Unsteady Vortex Lattice Method.

mesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{GeometryBasics.Point3{T}, 1})}

ringMesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{GeometryBasics.Point3{T}, 1})}

wakeMesh::GeometryBasics.Mesh{3, T, FT, Names, VAT} where {T, FT<:GeometryBasics.AbstractFace, Names, VAT<:(Tuple{var"#s13", Vararg{Union{AbstractVector{T}, GeometryBasics.FaceView{T, AVT} where AVT<:AbstractVector{T}} where T}} where var"#s13"<:AbstractArray{GeometryBasics.Point3{T}, 1})}

sizes::Sizes

wakeSizes::Sizes

panelProperties::PanelProperties

modelProperties::AeroModelProperties

segmentProps::SegmentProperties

K8::SparseArrays.SparseMatrixCSC{T, Int64} where T

K9::SparseArrays.SparseMatrixCSC{T, Int64} where T

L3::Matrix

L4::Matrix

L5::Matrix

L6::Matrix

L7::Matrix

L8::Matrix

L9::SparseArrays.SparseMatrixCSC{T, Int64} where T

L10::SparseArrays.SparseMatrixCSC{T, Int64} where T

AeroPanels.AeroModelProperties — Type

struct AeroModelProperties{T}

Aerodynamic model properties.

c::Any

b::Any

S::Any

symmXZ::Bool

symmXY::Bool

bodyFixedCS::StaticArraysCore.SMatrix{3, 3, T, 9} where T

AeroPanels.PanelProperties — Type

struct PanelProperties{T}

Properties of the panels in the aerodynamic mesh.

normal::Array{GeometryBasics.Point3{T}, 1} where T

rMid::Array{GeometryBasics.Point3{T}, 1} where T

rCollocation::Array{GeometryBasics.Point3{T}, 1} where T

rControl::Array{GeometryBasics.Point3{T}, 1} where T

area::Vector

span::Vector

Solvers

AeroPanels.AeroSolve — Function

AeroSolve(vb, model)
AeroSolve(vb, model, ρ)

Solve the steady aerodynamic problem for a given body velocity vb. Returns a SteadySolution.

AeroSolve(V, α, model)
AeroSolve(V, α, model, ρ)

Solve the steady aerodynamic problem for a given speed V and angle of attack α (degrees).

AeroSolve(vb, aeroModel)
AeroSolve(vb, aeroModel, ρ)

Solve for the steady-state aerodynamic solution of an unsteady model.

AeroPanels.SolveForces — Function

SolveForces(Γw, vb, dvb, model)
SolveForces(Γw, vb, dvb, model, ρ)

Return total aerodynamic forces (Steady + Unsteady) for given states and kinematics.

AeroPanels.SolveCirculation — Function

SolveCirculation(Γw, vb, model)

Calculate the time derivative of the wake circulation states (dΓw/dt).

AeroPanels.NormalWash — Function

NormalWash(vb, model)

Calculate and return the normal wash vector b.

AeroPanels.NormalWash! — Function

NormalWash!(b, vb, model)

Calculate the normal wash on each panel in-place into vector b.

AeroPanels.Circulation — Function

Circulation(b, model)

Solve for and return (Γp, Γw, Γs).

AeroPanels.Circulation! — Function

Circulation!(Γp, Γw, Γs, b, model)

Solve for panel, wake, and segment circulations in-place.

AeroPanels.AerodynamicForce — Function

AerodynamicForce(Γp, Γw, Γs, vb, model; ρ)

Calculate and return the aerodynamic force vector Fa on segments.

AeroPanels.AerodynamicForce! — Function

AerodynamicForce!(Fa, Γp, Γw, Γs, vb, model; ρ)

Calculate aerodynamic forces on segments in-place into vector Fa.

AeroPanels.SteadySolution — Type

struct SteadySolution{T}

The solution to a steady aerodynamic problem.

forceBody::GeometryBasics.Point3

forceStability::GeometryBasics.Point3

forceUnsteady::GeometryBasics.Point3

coefficientBody::GeometryBasics.Point3

coefficientStability::GeometryBasics.Point3

forceVecSeg::Array{GeometryBasics.Point3{T}, 1} where T

CL::Any

CD::Any

Secondary Functions

AeroPanels.AeroSurface2D — Type

struct AeroSurface2D{T<:Real} <: AeroSurface{T<:Real}

A 2D aerodynamic surface defined by a mesh of points in 3D space.

X::Matrix{T} where T<:Real

Y::Matrix{T} where T<:Real

Z::Matrix{T} where T<:Real

nc::Int64

ns::Int64

AeroPanels.Mirror — Function

Mirror(surface, axis)

Mirror an AeroSurface2D along a given axis (1=X, 2=Y, 3=Z).

AeroPanels.SegmentProperties — Type

struct SegmentProperties{T}

Properties and mapping indices for panel segments.

mid::Array{GeometryBasics.Point3{T}, 1} where T

r::Array{GeometryBasics.Point3{T}, 1} where T

aic3Ring::Array{GeometryBasics.Point3{T}, 2} where T

aic3Wake::Array{GeometryBasics.Point3{T}, 2} where T

indPosSeg::Vector{Int64}

indPosPanel::Vector{Int64}

indNegSeg::Vector{Int64}

indNegPanel::Vector{Int64}

nSpanSegments::Int64

nChordSegments::Int64

nTotalSegments::Int64

AeroPanels.SegmentCirculation — Function

SegmentCirculation(Γp, sp)

Return the segment circulation vector.

AeroPanels.SegmentCirculation! — Function

SegmentCirculation!(Γseg, Γp, sp)

Calculate segment circulations from panel circulations in-place.

AeroPanels.ProcessSegments — Function

ProcessSegments(meshRing, sizesRing, meshWake, sizesWake, symmXZ)

Initialize and calculate all segment properties.

AeroPanels.NumberOfStates — Function

NumberOfStates(model)

Return the number of states in the unsteady state-space model.

AeroPanels.GetFullWakeVector — Function

GetFullWakeVector(Γw, b, model)

Calculate the full wake circulation vector (including the Kutta edge) from transport wake states.

AeroPanels.UnsteadyPanelForces — Function

UnsteadyPanelForces(Γw, b, ρ, model, db)

Calculate the unsteady contribution to aerodynamic forces based on circulation time derivatives.

AeroPanels.SteadyForce — Function

SteadyForce(Γw, b, vb, ρ, model)

Calculate the quasi-steady portion of the aerodynamic forces.

AeroPanels.SolveSteadyCirculation — Function

SolveSteadyCirculation(b, model)

Solve for the steady-state wake circulation given normal wash b.

AeroPanels.UnsteadyWakeInfluence — Function

UnsteadyWakeInfluence(
    rCollocation,
    normal,
    ringMesh,
    wakeMesh,
    bodySizes,
    wakeSizes,
    symmXZ,
    Δt
)

Calculate the unsteady aerodynamic influence coefficients and state-space matrices. Based on Binder (2017) and AIAA (2018) publications.

Indexing and Utility

AeroPanels.Sizes — Type

struct Sizes

Struct to store the sizes and pre-calculated indices of the mesh elements.

sizes::Vector{Tuple{Int64, Int64}}

nSurfaces::Int64

totalPanels::Int64

totalVertices::Int64

totalSpanSegments::Int64

totalChordSegments::Int64

panelIndices::Vector{Matrix{Int64}}

vertexIndices::Vector{Matrix{Int64}}

spanSegmentIndices::Vector{Matrix{Int64}}

chordSegmentIndices::Vector{Matrix{Int64}}

AeroPanels.TEPanelIndex — Function

TEPanelIndex(sizes::Sizes)

Return a vector of global indices for all panels along the trailing edges.

AeroPanels.LEPanelIndex — Function

LEPanelIndex(sizes::Sizes)

Return a vector of global indices for all panels along the leading edges.

AeroPanels.PanelIndex — Function

PanelIndex(s::Int, i::Int, j::Int, sizes::Sizes)

Return the global index of the panel (i, j) on surface s.